requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
python-dotenv==1.0.0
mysql-connector-python==8.4.0
bcrypt==4.0.1
opencv-python==4.10.0.84
numpy==1.26.4
ultralytics==8.2.103
mediapipe==0.10.14
joblib==1.4.2
google-generativeai==0.8.3